40_ml_next/requirements.txt
fastapi
uvicorn
joblib
scikit-learn
numpy
pandas
boto3
python-dotenv